+++ src/Models/return/ReturnOrderCollection.ts
@@ -1,0 +1,7 @@
+import ReturnOrderCollectionInterface from '../../Interfaces/return/ReturnOrderCollection';
+
+export default class ReturnOrderCollection {
+    constructor(parameters: ReturnOrderCollectionInterface) {
+        Object.assign(this, parameters);
+    }
+}