+++ CHANGELOG.md
@@ -6,8 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+## [0.0.87] - 2022-03-11
 ### 🏡 Features
 - Add User types history resource items ([#412](https://github.com/whirli/whirli-client/pull/412))
+- Add user referral properties to user interface ([#417](https://github.com/whirli/whirli-client/pull/417))
+
+## [0.0.86] - 2022-03-03
+### Features
+- add return collection and shipping manifest endpoints ([#415](https://github.com/whirli/whirli-client/pull/415))
 
 ## [0.0.85] - 2022-02-21
 ### 🏡 Features